Add Installation page and update Browser Support
A new Installation index page has been added to the docs, advising on how to install FreshJuice on HubSpot websites. An update has also been made to the docs with details on browser support, specifying compatibility with the latest stable versions of Chrome, Firefox, Edge, and Safari and the lack of support for IE versions.