Update package dependencies and .nvmrc file
This commit performs several updates. It upgrades the postcss and tailwind-config-viewer versions to their latest releases in the package.json file. Additionally, it adds the sharp library to the list of dependencies. In the .nvmrc file, the lack of a newline at the end of the file has been corrected.