Add script to generate Open Graph images
This commit introduces a script, 'generate-og-images.js', that creates Open Graph (OG) images for posts. It auto-generates the images based on a set of categories (blogs, docs, pages) drawn from a JSON data file. Background colors for the images are also selected randomly from a defined set. The title and category are included in each image.